fix query tests
