+++ test/react-web/client/graphql/queries-1.test.tsx
@@ -94,190 +94,7 @@ describe('queries', () => {
     renderer.create(<ProviderMock client={client}><Container /></ProviderMock>);
   });
 
-  // XXX fix test
-  xit('correctly rebuilds props on remount', (done) => {
-    const query = gql`query pollingPeople { allPeople(first: 1) { people { name } } }`;
-    const data = { allPeople: { people: [ { name: 'Darth Skywalker' } ] } };
-    const networkInterface = mockNetworkInterface(
-      { request: { query }, result: { data }, newData: () => ({
-        data: {
-          allPeople: { people: [ { name: `Darth Skywalker - ${Math.random()}` } ] },
-        }
-      }) }
-    );
-    const client = new ApolloClient({ networkInterface });
-    let wrapper, app, count = 0;
-
-    @graphql(query, { options: { pollInterval: 10 }})
-    class Container extends React.Component<any, any> {
-      componentWillReceiveProps(props) {
-        if (count === 1) { // has data
-          wrapper.unmount();
-          wrapper = renderer.create(app);
-        }
-
-        if (count === 10) {
-          wrapper.unmount();
-          done();
-        }
-        count++;
-      }
-      render() {
-        return null;
-      }
-    };
-
-    app = <ProviderMock client={client}><Container /></ProviderMock>;
-
-    wrapper = renderer.create(app);
-  });
-
-  // XXX fix test
-  xit('correctly sets loading state on remounted forcefetch', (done) => {
-    const query = gql`query pollingPeople { allPeople(first: 1) { people { name } } }`;
-    const data = { allPeople: { people: [ { name: 'Darth Skywalker' } ] } };
-    const networkInterface = mockNetworkInterface(
-      { request: { query }, result: { data }, delay: 10, newData: () => ({
-        data: {
-          allPeople: { people: [ { name: `Darth Skywalker - ${Math.random()}` } ] },
-        },
-      }) }
-    );
-    const client = new ApolloClient({ networkInterface });
-    let wrapper, app, count = 0;
-
-    @graphql(query, { options: { forceFetch: true }})
-    class Container extends React.Component<any, any> {
-      componentWillMount() {
-        if (count === 1) {
-          expect(this.props.data.loading).toBe(true); // on remount
-          count++;
-        }
-      }
-      componentWillReceiveProps(props) {
-        if (count === 0) { // has data
-          wrapper.unmount();
-          setTimeout(() => {
-            wrapper = renderer.create(app);
-          }, 5);
-        }
-
-        if (count === 2) {
-          // remounted data after fetch
-          expect(props.data.loading).toBe(false);
-          expect(props.data.allPeople).toBeTruthy();;
-          done();
-        }
-        count++;
-      }
-      render() {
-        return null;
-      }
-    };
-
-    app = <ProviderMock client={client}><Container /></ProviderMock>;
-
-    wrapper = renderer.create(app);
-  });
-
-  // XXX fix test
-  xit('correctly sets loading state on remounted component with changed variables', (done) => {
-    const query = gql`
-      query remount($first: Int) { allPeople(first: $first) { people { name } } }
-    `;
-    const data = { allPeople: null };
-    const variables = { first: 1 };
-    const variables2 = { first: 2 };
-    const networkInterface = mockNetworkInterface(
-      { request: { query, variables }, result: { data }, delay: 10 },
-      { request: { query, variables: variables2 }, result: { data }, delay: 10 }
-    );
-    const client = new ApolloClient({ networkInterface });
-    let wrapper, render, count = 0;
-
-    @graphql(query, { options: ({ first }) => ({ variables: { first }})})
-    class Container extends React.Component<any, any> {
-      componentWillMount() {
-        if (count === 1) {
-          expect(this.props.data.loading).toBe(true); // on remount
-          count++;
-        }
-      }
-      componentWillReceiveProps(props) {
-        if (count === 0) { // has data
-          wrapper.unmount();
-          setTimeout(() => {
-            wrapper = renderer.create(render(2));
-          }, 5);
-        }
-
-        if (count === 2) {
-          // remounted data after fetch
-          expect(props.data.loading).toBe(false);
-          done();
-        }
-        count++;
-      }
-      render() {
-        return null;
-      }
-    };
-
-    render = (first) => (
-      <ProviderMock client={client}><Container first={first} /></ProviderMock>
-    );
-
-    wrapper = renderer.create(render(1));
-  });
-
-  // XXX fix test
-  xit('correctly sets loading state on remounted component with changed variables (alt)', (done) => {
-    const query = gql`
-      query remount($name: String) { allPeople(name: $name) { people { name } } }
-    `;
-    const data = { allPeople: null };
-    const variables = { name: 'does-not-exist' };
-    const variables2 = { name: 'nothing-either' };
-    const networkInterface = mockNetworkInterface(
-      { request: { query, variables }, result: { data }, delay: 10 },
-      { request: { query, variables: variables2 }, result: { data }, delay: 10 }
-    );
-    const client = new ApolloClient({ networkInterface });
-    let count = 0;
-
-    @graphql(query)
-    class Container extends React.Component<any, any> {
-      render() {
-        const { loading } = this.props.data;
-        if (count === 0) expect(loading).toBe(true);
-        if (count === 1) expect(loading).toBe(false);
-        if (count === 2) expect(loading).toBe(true);
-        if (count === 3) {
-          expect(loading).toBe(false);
-          done();
-        }
-        count ++;
-        return null;
-      }
-    };
-    const main = document.createElement('DIV');
-    main.id = 'main';
-    document.body.appendChild(main);
-
-    /* const render = (props) => { */
-    /*   ReactDOM.render(( */
-    /*     <ProviderMock client={client}> */
-    /*       <Container {...props} /> */
-    /*     </ProviderMock> */
-    /*   ), document.getElementById('main')); */
-    /* }; */
-
-    /* // Initial render. */
-    /* render(variables); */
-
-    /* // Prop update: fetch. */
-    /* setTimeout(() => render(variables2), 1000); */
-  });
+  // XXX reinsert `queries-2.test.tsx` after react 15.4
 
   it('correctly sets loading state on component with changed variables and unchanged result', (done) => {
      const query = gql`
@@ -1152,8 +969,7 @@ describe('queries', () => {
       </ProviderMock>);
   });
 
-  // XXX fix test
-  xit('exposes updateQuery as part of the props api', (done) => {
+  it('exposes updateQuery as part of the props api', (done) => {
     const query = gql`query people { allPeople(first: 1) { people { name } } }`;
     const data = { allPeople: { people: [ { name: 'Luke Skywalker' } ] } };
     const networkInterface = mockNetworkInterface({ request: { query }, result: { data } });
@@ -1164,8 +980,11 @@ describe('queries', () => {
       componentWillReceiveProps({ data }) { // tslint:disable-line
         expect(data.updateQuery).toBeTruthy();
         expect(data.updateQuery instanceof Function).toBe(true);
-        expect(data.updateQuery).not.toThrow();
-        done();
+        try {
+          data.updateQuery(() => done());
+        } catch (error) {
+          // fail
+        }
       }
       render() {
         return null;
